add database not done with database yet
