+++ app/screens/creategroup.tsx
@@ -1,0 +1,20 @@
+import { View, Text, Button } from 'react-native'
+import React from 'react'
+
+const Creategroup = () => {
+
+  const handleCreateGroup = () => {
+
+  }
+
+  return (
+    <>
+    <View>
+      <Text>Creategroup</Text>
+    </View>
+    <Button title='Create Group'/>
+    </>
+  )
+}
+
+export default Creategroup
