+++ package.json
@@ -17,10 +17,14 @@
   "dependencies": {
     "@expo/vector-icons": "^14.0.2",
     "@react-native-firebase/app": "19.0.0",
+    "@react-native-firebase/auth": "19.0.0",
+    "@react-native-firebase/database": "19.0.0",
+    "@react-native-firebase/firestore": "19.0.0",
     "@react-navigation/native": "^6.0.2",
     "expo": "^52.0.17",
     "expo-build-properties": "~0.13.1",
     "expo-constants": "~17.0.3",
+    "expo-dev-client": "~5.0.5",
     "expo-font": "~13.0.1",
     "expo-linking": "~7.0.3",
     "expo-router": "~4.0.11",
@@ -35,9 +39,7 @@
     "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.1.0",
-    "react-native-web": "~0.19.10",
-    "@react-native-firebase/auth": "19.0.0",
-    "expo-dev-client": "~5.0.5"
+    "react-native-web": "~0.19.10"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
