+++ tsconfig.json
@@ -11,6 +11,7 @@
   "include": [
     "**/*.ts",
     "**/*.tsx",
+    "*.tsx",
     ".expo/types/**/*.ts",
     "expo-env.d.ts"
   ]
